Move DCommand from commands to lib
